Branch for testing
This branch is for unit testing critical functions.
Testing framework is Jasmine.
Tests are contained in jasmine\spec
Tested functions are in jasmine\src
To see test results open SpecRunner.html
If more files are added to be tested/as tests, they must be included in
SpecRunner.html at the required points